dead takers now notifies other takers inside :take
We name the client the "dead taker" if it closed connection to Tarantool
during xq:take(). Tarantool closes socket and calls callback box.session.on_disconnect
and xqueue registers that taker inside this session can't consume tasks anymore.

But, if clients sleeps for too long on the xq:take() (ex. 60 seconds) then
next taker will receive task only when new task will be produced. It is too long
to wait, so this fix implements alarm of the other takers (if there are any) in
"dead taker"