locking mechanism
This commit adds new feature to xqueue - `lockable`.
User may specify a `lock` field in fields map when doing
xqueue.upgrade(). If specified xqueue checks that there is an additional
index with fields (lock, status, ...). This index is used by put(),
ack(), bury() and kill() methods in order to determine if current task
should be locked or unlocked.
This feature allows to process tasks only in a strict order or in other
words, doesn't allow workers to take a task when another task by lock
field is already being processed.

Notable additions:
* put(): When putting a task additional 1-3 queries are done to find if
  there is any other task in the queue with statuses L, T or R. If there
  is at least one task, then current task's status will be set to L
  rather than R.
* ack(), kill(), bury(): In these methods after successful status change
  of task xqueue will try to find a task with L status within `lock`
  index and wake it up, transferring to an R state.

This `lockable` feature currently cannot be used alongside with ttl or delayed features.